Instead of meaningful commit messages I'll just use this as a blog.
The game as-is sucks. There are only two persons who tried it. The first
one only had the code challenges and the second one had to go through
all three riddles, none of which he succeded - and those are
programmers, not normal people.

So far, I've failed completely at conveying what the game is about, at
guiding people to solve riddles (if you get stuck, you're stuck, there's
no hotline or anything - not even hints).

Part of me just wants to let it like that, parts of me wants to try and
get feedback from the interweb, but I'm pretty sure that getting stuck
on the first riddle and never ever coming back to it will be everyone's
reaction. Which sucks.

So, yeah, whatever.