More inline blogging...
I got some pretty solid feedback from @snipercommand on Twitter.
Recording the logs of people's playthrough would be interesting indeed,
but I did not want to add a dynamic component to the server side for
this game - it takes effort to keep something running. So, I'll see
if I can get something as a Linux service so that it survives
server reboots (which are not frequent, but they happen.)

I've added hints. It looks a bit messy but at least, it will help people
who get stuck. I have a feeling the game is too easy with it now, but
then gain if you're stuck indefinitely it's no fun either. So, there.