More inline blogging..
I realized the deadline was friday - I knew it was coming but I didn't
know it was so soon.

I tweaked some parts of the dialog, to make them shorter and hopefully
more striking - it's hard to tell. I tend to think that the whole dialog
sucks but for some reason some people told in their feedback that it
really set the stage for them.

Perhaps graphics or music would help, or on the contrary the absence of
them helps understanding that you are dealing with an entity that
clearly does not care about you, or your comfort, or your fate.

I think there's enough build-up for the two-guards problem now, which
is the next step. We'll see how this pans out.

I also have to worry about cache busting because Cloudflare / browser
caches are playing strange tricks on people, having stuff undefined
which makes the game impossible to play even if you're not a dumb nut.

I took a look at Middleman, the static site generator, but it seems
perhaps overkill for just a game like that. Besides, more ruby slowness
= grrr. We'll see what I do eventually.