Moar inline blogging.
I now have a good idea for difficulty settings, but that will have to
wait for a next version.

ace is now hosted locally, and I'm showing the gutter again, so that
syntax error are shown as the player is typing.

The instructions are more explicit, and I'm giving more context for the
questions, along with more belittling to set the mood just right.

The build process is slightly altered, with the javascript being
minified.

The game now forces players to use only the keyboard after the first few
cards. There are several reasons for that - 1) it's the first instance
of belittling that happens, which is good for the atmosphere, 2) it
forces players into a more passive role, having hands on the keyboard at
all time, 3) it's just more convenient and allows the game to flow more
smoothly.

I think the path leading up to the lying guards problem is now smooth
enough that someone could go through it without any external help. I'll
try some user testing while restraining myself from any comments and see
what happens.