Numannnnnnnnnnnns
I introduced the idea of vehicles, boarding, making numans cross
a chasm.

The human player has no idea why though - at this point he/she is
just following instructions. Which is just splendid, really...

I shall introduce a new voice as soon as possible.

As for infrastructure, there's a system of extensions now - which
is useful if you don't want to bloat the eval'd code, but yet share code
between different questions instead of entering the realm of copypasta.